requirements.txt
langextract>=1.1.0
google-generativeai>=0.3.0
langgraph>=0.0.1
pdfplumber>=0.9.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
